+++ test/classlist/test_operation.rb
@@ -4,6 +4,7 @@ require "test_helper"
 
 require "classlist/add"
 require "classlist/remove"
+require "classlist/reset"
 
 class TestClasslistOperation < Minitest::Test
   def test_add
@@ -90,6 +91,22 @@ class TestClasslistOperation < Minitest::Test
     result = base + change
 
     assert_equal(["this"], result.to_a)
+  end
+
+  def test_adding_after_removing
+    base = Classlist.new("foo")
+    removal = Classlist::Remove.new("foo")
+    addition = Classlist::Add.new("addition")
+    result = base + removal + addition
+    assert_equal(["addition"], result.to_a)
+  end
+
+  def test_adding_a_classlist_instance_assumes_add
+    base = Classlist.new("foo")
+    removal = Classlist::Remove.new("foo")
+    addition = Classlist.new("addition")
+    result = base + removal + addition
+    assert_equal(["addition"], result.to_a)
   end
 
   def test_all_the_operations